Re worked Collapse
Collapse is now a Tagged Enum, and we work off
of a subset of Collapsed - ResolvedCollapse - when
we have executed an Effect.

The two ReEffects, Sync and Promise, now have slightly different
return shapes. These are more strongly typed, and the pattern matching
guides users to handle all cases necessary.

The useSync matcher cannot fall into a pending state, as it is
synchronously executed, so only the success and error cases are
required.

The usePromise matcher requires all cases to be handled.

The return types of all matcher functions are unioned together propery -
whew!

Some helper types have bubbled up into a type-utils function.

Simplified usePromise quite a bit.